# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KA Imports API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  promoPrice?: number | null;
  image: string;
  category: string;
}

export interface CartItem {
  id: string;
  name: string;
  quantity: number;
  price: number;
  image?: string;
}

export interface ClientInfo {
  name: string;
  email: string;
  phone: string;
  document: string;
}

export interface AddressInfo {
  cep: string;
  street: string;
  number: string;
  complement?: string;
  neighborhood?: string;
  city: string;
  state: string;
}

export type GeneratePixRequestShippingType = string;

export interface GeneratePixRequest {
  client: ClientInfo;
  products: CartItem[];
  amount: number;
  shippingType: GeneratePixRequestShippingType;
  includeInsurance: boolean;
  orderId: string;
}

export interface PixResponse {
  transactionId: string;
  status: string;
  pixCode: string;
  pixBase64: string;
  pixImage?: string;
  expiresAt: string;
  orderId: string;
  receiptUrl?: string | null;
}

export interface PixStatusResponse {
  transactionId: string;
  status: string;
  paidAt?: string | null;
}

export type CreateOrderRequestShippingType = string;

export interface CreateOrderRequest {
  client: ClientInfo;
  address?: AddressInfo;
  products: CartItem[];
  shippingType: CreateOrderRequestShippingType;
  includeInsurance: boolean;
  subtotal: number;
  shippingCost: number;
  insuranceAmount: number;
  total: number;
  paymentMethod?: string;
  cardInstallments?: number;
}

export interface Order {
  id: string;
  client: ClientInfo;
  address?: AddressInfo;
  products: CartItem[];
  shippingType: string;
  includeInsurance: boolean;
  subtotal: number;
  shippingCost: number;
  insuranceAmount: number;
  total: number;
  status: string;
  createdAt: string;
}

export interface AdminOrder {
  id: string;
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  clientDocument: string;
  addressCep?: string | null;
  addressStreet?: string | null;
  addressNumber?: string | null;
  addressComplement?: string | null;
  addressNeighborhood?: string | null;
  addressCity?: string | null;
  addressState?: string | null;
  products: CartItem[];
  shippingType: string;
  includeInsurance: boolean;
  subtotal: number;
  shippingCost: number;
  insuranceAmount: number;
  total: number;
  status: string;
  paymentMethod?: string;
  cardInstallments?: number | null;
  proofUrl?: string | null;
  transactionId?: string | null;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetProducts200 = {
  products: Product[];
  categories: string[];
};

export type PixCallbackBody = { transactionId?: string; status?: string; [key: string]: unknown };

export type PixCallback200 = {
  ok: boolean;
};

export type GetAdminOrders200 = {
  orders: AdminOrder[];
};