# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KA Imports API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all products from Google Sheets
 */
export const GetProductsResponse = zod.object({
  products: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
      price: zod.number(),
      promoPrice: zod.number().nullish(),
      image: zod.string(),
      category: zod.string(),
    }),
  ),
  categories: zod.array(zod.string()),
});

/**
 * @summary Generate a PIX payment
 */
export const GeneratePixBody = zod.object({
  client: zod.object({
    name: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    document: zod.string(),
  }),
  products: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
    }),
  ),
  amount: zod.number(),
  shippingType: zod.enum(["normal", "express"]),
  includeInsurance: zod.boolean(),
  orderId: zod.string(),
});

export const GeneratePixResponse = zod.object({
  transactionId: zod.string(),
  status: zod.string(),
  pixCode: zod.string(),
  pixBase64: zod.string(),
  pixImage: zod.string().optional(),
  expiresAt: zod.string(),
  orderId: zod.string(),
  receiptUrl: zod.string().nullish(),
});

/**
 * @summary Check PIX payment status
 */
export const GetPixStatusParams = zod.object({
  transactionId: zod.coerce.string(),
});

export const GetPixStatusResponse = zod.object({
  transactionId: zod.string(),
  status: zod.string(),
  paidAt: zod.string().nullish(),
});

/**
 * @summary PIX payment callback from gateway
 */
export const PixCallbackParams = zod.object({
  token: zod.coerce.string(),
});

export const PixCallbackBody = zod.record(zod.string(), zod.unknown());

export const PixCallbackResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  client: zod.object({
    name: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    document: zod.string(),
  }),
  products: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
    }),
  ),
  shippingType: zod.enum(["normal", "express"]),
  includeInsurance: zod.boolean(),
  subtotal: zod.number(),
  shippingCost: zod.number(),
  insuranceAmount: zod.number(),
  total: zod.number(),
});